cleaned up timer code slightly
git-svn-id: https://mspsim.svn.sourceforge.net/svnroot/mspsim/mspsim@297 23d1a52b-0c3c-0410-b72d-8f29ab48fe35

+++ se/sics/mspsim/core/MSP430.java
@@ -280,9 +280,12 @@ public class MSP430 extends MSP430Core {
   }
 
   private void printCPUSpeed(int pc) {
+    // Passed time
     int td = (int)(System.currentTimeMillis() - time);
-    int cd = (int) (cycles - lastCycles);
-    int cpud = (int) (cpuCycles - lastCpuCycles);
+    // Passed total cycles
+    long cd = (cycles - lastCycles);
+    // Passed "active" CPU cycles
+    long cpud = (cpuCycles - lastCpuCycles);
 
     if (td == 0 || cd == 0) {
       return;
@@ -295,7 +298,7 @@ public class MSP430 extends MSP430Core {
 			 + 1000 * (cpud / td ) + " cyc/s  "
 			 + (10000 * cpud / cd)/100.0 + "%");
     }
-    lastCPUPercent = (10000 * cpud / cd)/100.0;
+    lastCPUPercent = (10000 * cpud / cd) / 100.0;
     time = System.currentTimeMillis();
     lastCycles = cycles;
     lastCpuCycles = cpuCycles;
